Moved unsaved-indicator
